+++ assets/appFramework/angular/test/e2e/src/e2eTestRunner/webdriverio/sourceLanguage/javascript/jasmine/app.spec.js
@@ -18,9 +18,20 @@ describe("App", () => {
             .waitForText("#root > ng-component > div", 20000)
             .element("#root > ng-component > div")
             .getText()
-                .then((rootContent) => {
-                    expect(rootContent).toEqual("Hello UniteJS World!");
-                });
+            .then((rootContent) => {
+                expect(rootContent).toEqual("Hello UniteJS World!");
+            });
+    });
+
+    it("the font size is set", () => {
+        return browser
+            .url("/")
+            .waitForText("#root > ng-component > div", 20000)
+            .element(".child-style")
+            .getCssProperty("font-size")
+            .then((fontSize) => {
+                expect(fontSize.value).toEqual("20px");
+            });
     });
 });