+++ assets/appFramework/aurelia/test/e2e/src/e2eTestRunner/protractor/sourceLanguage/javascript/mochachai/app.spec.js
@@ -1,7 +1,7 @@
 /**
  * Tests for App.
  */
-import {expect} from "chai";
+import { expect } from "chai";
 
 describe("App", () => {
     it("the title is set", (done) => {
@@ -16,12 +16,23 @@ describe("App", () => {
             });
     });
 
-    it("the router text is set", (done) => {
+    it("the root text is set", (done) => {
         browser.loadAndWaitForAureliaPage("/")
             .then(() => {
                 $("router-view").getText()
                     .then((routerContent) => {
                         expect(routerContent).to.equal("Hello UniteJS World!");
+                        done();
+                    });
+            });
+    });
+
+    it("the font size is set", (done) => {
+        browser.loadAndWaitForAureliaPage("/")
+            .then(() => {
+                $(".child-style").getCssValue("font-size")
+                    .then((fontSize) => {
+                        expect(fontSize).to.equal("20px");
                         done();
                     });
             });