+++ assets/appFramework/aurelia/test/e2e/src/e2eTestRunner/webdriverio/sourceLanguage/typescript/mochachai/app.spec.ts
@@ -15,13 +15,23 @@ describe("App", () => {
             });
     });
 
-    it("the router text is set", () => {
+    it("the root text is set", () => {
         return browser
             .loadAndWaitForAureliaPage("/")
             .element("router-view")
             .getText()
             .then((routerContent) => {
                 expect(routerContent).to.equal("Hello UniteJS World!");
+            });
+    });
+
+    it("the font size is set", () => {
+        return browser
+            .loadAndWaitForAureliaPage("/")
+            .element(".child-style")
+            .getCssProperty("font-size")
+            .then((fontSize) => {
+                expect(fontSize.value).to.equal("20px");
             });
     });
 });