+++ assets/appFramework/react/generate/component/src/{GEN_NAME_SNAKE}.tsx
@@ -2,13 +2,10 @@
  * {GEN_NAME_HUMAN} component class.
  */
 import * as React from "react";
-
-const {GEN_NAME_CAMEL}Style = {
-    fontSize: "20px"
-};
+import "./{GEN_NAME_SNAKE}.css";
 
 export class {GEN_NAME_PASCAL} extends React.Component {
     public render(): JSX.Element {
-        return (<span style={{GEN_NAME_CAMEL}Style}>Hello {GEN_NAME_HUMAN}</span>);
+        return (<span className="{GEN_NAME_SNAKE}">Hello {GEN_NAME_HUMAN}</span>);
     }
 }